tests/wpt/web-platform-tests/xhr/formdata/has-formelement.html

<!doctype html>
<html lang=en>
<meta charset=utf-8>
<title>FormData: has (with form element)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="https://xhr.spec.whatwg.org/#dom-formdata-get" />
    <link rel="help" href="https://xhr.spec.whatwg.org/#dom-formdata-getall" />
<div id="log"></div>
<form id="form">
    <input type="hidden" name="key" value="value1">
    <input type="hidden" name="key" value="value2">
</form>
<form id="empty-form"></form>
<script>
    test(function() {
        assert_equals(new FormData(document.getElementById('form')).has('key'), true);
    }, 'testFormDataHasFromForm');
    test(function() {
        assert_equals(new FormData(document.getElementById('form')).has('nil'), false);
    }, 'testFormDataHasFromFormNull');
    test(function() {
        assert_equals(new FormData(document.getElementById('empty-form')).has('key'), false);
    }, 'testFormDataHasFromEmptyForm');
</script>
